requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
pyjwt
cachetools
mysql-connector-python
